NOT WORKING - This branch contains work to support client side localization so that asset-manager would provide localized version's of the different js files that you create.
Did not finish this implementation because of time limitation and because
the solution was not feeling very elegant with respect to the problem.
The file path resolution space was getting really muddy when you mixed the
locales with the minification since we are also using inject to load the files
client side and we'd have to serialize all of that info to the client.